Refactor StartModal into more general game setup modal
- Move into it's own directory with renamed component
- Move out singleplayer timer setup into it's own component
